+++ src/main/webapp/js/search_bar.js
@@ -1,0 +1,21 @@
+/**
+ * Toggle the search bar appearance based on event (To be moved to other file)
+ * @param obj - the button invoking the click
+ * @param evt - the click event
+ */
+function searchToggle(obj, evt){
+    var container = $(obj).closest('.search-wrapper');
+
+    if(!container.hasClass('active')){
+        container.addClass('active');
+        evt.preventDefault();
+    }
+    else if (container.hasClass('active') && $(obj).closest('.input-holder').length == 0){
+        container.removeClass('active');
+        // clear input
+        container.find('.search-input').val('');
+        // clear and hide result container when we press close
+        container.find('.result-container').fadeOut(100, function(){$(this).empty();});
+    }
+}
+
