+++ src/main/webapp/player.html
@@ -19,7 +19,12 @@
         <link rel="stylesheet" href="css/style.css">
         <link rel="stylesheet" href="css/search-bar.css">
         <link rel="stylesheet" href="css/player.css">
-        <script defer src="js/script.js"></script>
+        <script defer src="js/time_manipulation.js"></script>
+        <script defer src="js/search_bar.js"></script>
+        <script defer src="js/error_handling.js"></script>
+        <script defer src="js/yt_player.js"></script>
+        <script defer src="js/constants.js"></script>
+        <script defer src="js/captions.js"></script>
     </head>
     <body>
         <div class="container">
